+++ examples/scripts/spm_AdamW.py
@@ -57,7 +57,7 @@ optim = pybop.AdamW(
     cost,
     verbose=True,
     allow_infeasible_solutions=True,
-    sigma0=0.05,
+    sigma0=0.02,
     max_iterations=100,
     max_unchanged_iterations=20,
 )
@@ -77,4 +77,4 @@ pybop.plot_parameters(optim)
 
 # Plot the cost landscape with optimisation path
 bounds = np.asarray([[0.5, 0.8], [0.4, 0.7]])
-pybop.plot2d(optim, bounds=bounds, steps=15)
+pybop.plot_voronoi2d(optim, bounds=bounds)